QA12-026 Check menus generated for GPS plugin (transformations of Why3)
This checks that all transformations that are actually generated in the
Why3 plugins exist. The shared file to the GPS plugin should be
generated_menus.py. It is installed by the Makefile in
$(INSTALLED)/share/config/generated_menus.json

* Makefile
(why3): developper only line which fails if a transformation does not
exist. If it is successful, generated_menus.json should be installed and
used by the python plugin

* scripts/why3menus.py
This checks that all that is generated in ideal_transf_list actually
exists in why3 --list-transforms.

* generated_menus.json
The generated file which contains ideal_transf_list (used to generate
menus)

Change-Id: Ib7fddd4b89c6502a62558c52d9db2503e149c514