S924-018 only allow call to Pledge functions in positive contexts
Calls to functions annotated with the Pledge Annotate pragma are
translated differently for proof and for execution. As a result,
an assertion containing such a call can fail at runtime even if it is
proved by the tool. To avoid that, we restrict the context in which calls
to these functions can occur so that these calls appear positively either
directly inside a contract pragma or inside boolean operations or
conditional expressions nested in these pragmas.

* /gnat2why/spark/spark_definition.adb
(Mark_Call): Check the context of calls to functions with a Pledge
 annotation.

* tests/S415-012__bad_pledge_calls/test.out
Update test output.

* tests/S924-018__pledge_calls/
File test.

Change-Id: I57e389f05bd1786569c7b9fee0e9a83861737844