S925-004 Translate Compile_Time_Error pragmas as assumptions
Since the opposite of the property will be checked now at compilation,
we can assume it in proof.

* /gnat2why/why/gnat2why-expr.adb
(Transform_Pragma_Check): Translate Compile_Time_Error as an assumption
 of the negated property.

* tests/S925-004__compile_time_error/repro.ads
* tests/S925-004__compile_time_error/test.out
Assertion is proved now.

* tests/S925-004__compile_time_assert
New test showing a case where the property cannot be determined statically
and should lead to a compilation error.

Change-Id: I683ac6c6d7f00ccedd79040cf27887877fcdca0a