S920-003 add test to cover internal call to non-returning procedure
Before support for tasking we had explicit and global parameters; now we
also have implicit parameters (for internal calls within a protected unit).
This change was not reflected in all places in the code, especially where it
only affects corner-cases. This is a test with one such a corner-case.

* testsuite/gnatprove/tests/S920-003__flow_implicit_noret/
Add; we would expect a check about all paths in Caller never returning.

Change-Id: I9384c4457b1b5edbc41435c0be03879cb15957f2