S920-003 use execution kind for call vertices
In CFGs we used to connect vertices for subprogram calls using the "default"
edges and vertices for the call parameters using the "default" or "abnormal
termination" or "infinite execution edges". This resulted in inconsistent
handling of non-returning procedures with and without parameters.

Now we handle both kinds of vertices in the same way and get consistent
checks for calls to all non-returning procedures.

* gnat2why/flow/flow-control_flow_graph.adb
(Linkup): do not special-case subprogram call vertices.

* testsuite/gnatprove/tests/M611-010__no_return_illegal/test.out
* testsuite/gnatprove/tests/O521-002__no_return/test.out
* testsuite/gnatprove/tests/P602-007__protected_no_return/test.out
* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/test.out
New checks about all paths raising exceptions in non-terminating procedures
without a No_Return aspect; diffs in warning and check related to generated
global are because generating globals doesn't mix well with non-returning
procedures.

* testsuite/gnatprove/tests/S920-003__flow_noret/test.out
Spurious check about a potentially uninitialized output is gone.

Change-Id: Ie221bfeebcce32f1fdcfb8e7315e6551ff690786