R417-042 Update rules for Proof_In constants
Revert comment with the same TN, whose message reads:

"Proof_In constants do not participate in data- and information-flows, and
 as such do not belong to Global and Depends contracts. This is particularly
 important for constants with variable inputs that could be referred to from
 subtype predicates, so that the exact location of predicate checks can be
 ignored in flow analysis."

because we agreed that the ultimate solution is to handle implicit checks
for type predicates in the same way by both flow and proof (but sadly that
it is not easy to implement).

Note: we couldn't mechanically revert this commit, because rules in the
modified section has been renumbered.

* docs/lrm/source/subprograms.rst
Revert changes to verification rules regarding Global.

Change-Id: Id907f62af0316a8cdb8fb0f347544b108710374f