R417-042 deconstruct special-casing of type aspects at declaration
In flow we detected evaluation of Dynamic_Predicate and Type_Invariant at
object declarations, but not anywhere else; this was inconsistent and so is
deconstructed in this patch.

Now we only detect evaluation of Default_Initial_Condition at object
declarations, because it is the only place where it is evaluated

* gnat2why/flow/flow-control_flow_graph.adb
(Add_Vertex_For_DIC): renamed from Add_Vertex_For_Type_Aspect; cleaned-up to
 only support DIC.

Otherwise, remove code that now become dead and unreferenced.

* testsuite/gnatprove/tests/PB07-004__flow_calls_in_type_aspect/test.out
* testsuite/gnatprove/tests/Q512-013__flow_predicate_calls/test.out
We no longer detect potentially blocking operations executed when evaluating
Dynamic_Predicate at object declaration; this hole will be fixed once the
predefined Yield operation is annotated with non-null Global contract.

* testsuite/gnatprove/tests/KB29-037__subtype/test.out
New warning about an object that was only declared to for evaluation of
Dynamic_Predicate.

* testsuite/gnatprove/tests/R417-042__flow_type_aspects/test.out
The output is now consistent both for object declared with an initialization
expression and without.

Change-Id: Ib6326933b24b5dd4d24402b70b34c8132618ce6c