P609-046 down-project abstract states with unambiguous refinements
We now down-project abstract states if we can determine that their
refinement is unambiguous. Refinement is unambiguous in the following
cases:

1) If the caller can see that the abstract state (which appears as a
global Input/Proof_In) has just ONE constituent, then we know what
Redined_Global needs to be. If it has ZERO constituents, then we know
that it is a null refinement.

2) If the abstract state appears as a Global Output then this is also
enough to know what the Refined_Global needs to be. (i.e. all of its
constituents are of mode Output). In all other cases, the refinement is
ambiguous, so we can't work out what it needs to be, and thus we need to
generate the refinement.

* gnat2why/flow/flow_refinement.adb
* gnat2why/flow/flow_refinement.ads
(Mentions_State_With_Ambiguous_Refinement): Create new function that
determines the above condition. This replaces
Mentions_State_With_Visible_Refinement.

(Refinement_Needed): Make use of new function.

* gnat2why/flow/flow_utility.adb
(Get_Depends, Get Globals): Make use of new function.

* gnat2why/flow/flow-analysis.adb
(Emit_Check_Message): Add a condition such that we only emit a check
message if the use of an uninitialised Global is not 'Final in the CFG.

* testsuite/gnatprove/tests/Q821-003__flow_visible/test.opt
Exclude test from testsuite. This patch exposes a problematic
interaction between Flow_Refinement.Expand, which requires that the
entity be SPARK_Mode => On in order to check State_Refinement_Is_Visible
- and Mark_Subprogram_Specification which assumes SPARK_Mode => Off for
all entities when doing top-down marking.

* testsuite/gnatprove/tests/M227-029__flow_records_and_arrays/test.out
* testsuite/../M304-013__flow_check_contracts/test.out
* testsuite/../M314-012__flow_abstract_state/test.out
* testsuite/../M325-008__flow_contract_pragmas/test.out
* testsuite/../M503-001__typinv/test.out
* testsuite/../MA11-028__flow_compute_refined_global_depends/test.out
* testsuite/../N108-029__flow_generate_initializes_1/test.out
* testsuite/../N502-006__flow_justify_uninit_state/test.out
* testsuite/../O219-034__flow_global_checks/test.out
* testsuite/../O303-022__abstract_only/test.out
* testsuite/../O515-015__unused_global/test.out
* testsuite/../O515-017__generated_refined/test.out
* testsuite/../O918-007__generic_part_of/test.out
* testsuite/../OA20-032__refinement/test.out
* testsuite/../OC08-044__refinement/test.out
* testsuite/../P125-015__abstr_refine/test.out
* testsuite/../P527-025__flow_abstract_state/test.out
* testsuite/../P609-046__global/test.out
* testsuite/../P609-046__global_refined/test.out
* testsuite/../P615-019__bronze/test.out
* testsuite/../P820-002__protected_out/test.out
* testsuite/../P902-007__flow_spurious_global_mode_error/test.out
* testsuite/../PA25-031__flow_wrong_depends/test.out
* testsuite/../Q204-001__linted/test.out
* testsuite/../Q526-008__flow_child_refine/test.out
* testsuite/../Q810-019__flow_proof_in/test.out
* testsuite/../Q915-014__flow_useless_global/test.out
* testsuite/../R426-033__flow_depends/test.out
* testsuite/../R522-032__flow_libgfxinit/test.out
* testsuite/../R524-009__flow_no_global_generation_switch/test.out
* testsuite/../R524-009__flow_no_global_generation_switch_children/test.out
* testsuite/../R618-018__flow_global/test.out
* testsuite/../RA24-012__flow_incomplete_global/test.out
* testsuite/../S510-037__init_by_proof/test.out
* testsuite/../S530-019__init_by_proof/test.out
* testsuite/../TU__global_illegal/test.out
* testsuite/../TU__type_invariant__legal/test.out
* testsuite/../adacore_u/test.out
* testsuite/../nuclear_system/test.out
Update test output.

* testsuite/gnatprove/tests/nuclear_system/nuclear_reactors.ads
Minor change to justify check message.

* docs/ug/gnatprove_by_example/results/data_initialization.flow
* docs/ug/gnatprove_by_example/results/data_initialization.prove
* docs/ug/gnatprove_by_example/results/gen_refined_global.flow
* docs/ug/gnatprove_by_example/results/gen_refined_global.prove
Update user guide examples.

Change-Id: I882ff63c78814256d6670dcdcdf2607516086f12