S818-003 suppress warnings on paths with no effects
We already suppressed warnings on *calls* with no effects, like:

   procedure ... is
   begin
      Dummy_Call;
   end;

and with this patch we will also suppressed similar warning on *paths* with
no effects, like:

   procedure ... is
   begin
      if ... then
         Dummy_Call;
      end if;
   end;

* gnat2why/flow/flow-analysis.adb
(Is_Dummy_Call): lift from being nested in Find_Ineffective_Statements so
 that it can be reused in Find_Ineffective_Imports_And_Unused_Objects.

(Is_Final_Use): handle dummy calls as effects.
(Is_Final_Use_Any_Export): ditto.

* testsuite/gnatprove/tests/O914-009__limit_subp_for_entries/test.out
This update is not really intended, but it is harmless as well. It looks
like something is wrong for protected entries that merely test the entry
barrier but otherwise have no effects; this is clearly a corner-case, to
be investigated separately, if ever needed.

Change-Id: I92dc53c010139246e58487f5d5504c42b997c0ab