SA03-010 correct SPARK_Mode for parameter of predicate functions
For predicate functions, we do not use the SPARK_Mode of the enclosing
part but the SPARK_Mode of the type, as they are declared in the private
part even for public types. We now do the same thing for their parameters.

* gnat2why/spark/spark_definition.adb
(SPARK_Pragma_Of_Entity): For the formal parameter of a predicate
 function, use the SPARK_Mode of the predicate function.

* tests/SA03-010__formal_of_pred_fun/test.out
Remove spurious output.

Change-Id: I7bba5b819ab2536493ef4e6e6fcdb8e0d0e47697