deconstruct record discriminant for package bodies
In flow we used to distinguish between analysing a package spec alone and
package spec+body, but this was confusing, because there are really 3 parts
visible+private+body of a package.

Now we just have a package and all the rest it deduced by querying the
marking status.

(no-tn-check)

* gnat2why/flow/flow_types.ads
(Analyzed_Subject_Kind): remove Kind_Package_Body and propagate as needed.

Change-Id: Id81e827315991d3ad58437410436a4f2053dfd2d