R625-017 add tests with references to a discriminant
* testsuite/gnatprove/tests/R625-017__flow_discr/
For P, Global contract on Nested appears fine; flow check appears spurious
(strictly speaking, SPARK RM is not clear about this, but the current
behaviour in flow is more by accident than by design).

For Main, contracts for F1 and F2 seem acceptable, but certainly weird,
because are different for constrained and unconstrained object declaration.

Change-Id: I397c58414aaaed9cde5d15dbcf9bbc672dc86a10