SA10-013 add test for grouping of protected components
The expected output isn't terribly helpful; one need to look at the CFGs
to see the problem or add a postcondition to Dependency_Path saying that
the result is a non-empty sequence (which is the subject of SA02-015).

* testsuite/gnatprove/tests/SA10-013__flow_protected_group/
CFG produced for two Flip procedures (one protected and the other not
protected) should be similar, but they are not. Once fixed, we should get
some traces in test output.

Change-Id: Ibd7e0dd46f72647a1dd5a997e312b2a9d0c610e3