SA10-013 fix grouping of protected components in CFG
This patch fixes the grouping in CFG, so that graphs for protected and
non-protected operations in the test for this TN look same. Test outputs
are not yet affected, because PDGs for composite objects are still broken.

* gnat2why/flow/flow-control_flow_graph.adb
(Create_Record_Tree): do not special-case concurrent components.

* gnat2why/flow/flow_types.adb
* gnat2why/flow/flow_types.adb
(Parent_Record): same as above; this allows the contract to be strengthen.

* testsuite/gnatprove/tests/SA10-013__flow_protected_group/
CFGs now look as they should (but this is not tracked in the test output).

Change-Id: Ic514a67c3ac09d80630cd4ebb04bee576562a542