SA17-037 old values of borrowed objects not allowed in pledges
Inside calls to pledge functions, it is allowed to read the object
borrowed by the borrower the pledge refers to. Indeed, this object is
universally quantified in the pledge, so this is not an actual read.
However, if the object is below a 'Loop_Entry or 'Old attribute, then
it is not the quantified variable anymore, and reading it should be
forbidden.

* gnat2why/why/gnat2why-borrow_checker.adb
(Check_Not_Borrowed): Do not search for enclosing pledge function
 calls when the expression is below a 'Loop_Entry or 'Old attribute.

* tests/SA17-037__pledge
File test.

Change-Id: I0149b57985d452ce2dc6484e10ab3aceb913de17
(cherry picked from commit 10178525068e7fffde073e2e6ee204965f685b6d)