SA28-022 assume value of discriminants when checking components
When translating the declaration of a record type with discriminants, a
check may be emitted for the subtype associated to a component declaration.
We now assume the value of the discriminant for components declared
inside a variant part.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Wrap checks for record components inside
 conditional statement so that the discriminant is constrained for the
 checks.

* tests/M830-004__checks/test.out
Better counterexamples for discriminant dependant component.

* tests/PB02-004__inherited_sub/test.out
* tests/SA28-022__nest_discr/test.out
Range check proved for discriminant dependant component.

* tests/ghc_sort/test.out
Proof regression, could be regained with more steps.

Change-Id: I6fab8f96996ae07ee07544e0913cf76faeb7e7cf