SA16-067 do not compute length on bv on arrays indexed by signed types
Converting 'First and 'Last to bitvectors may be incorrect when computing
the length of an array indexed by signed integers. Only do the computation
on bitvector types if the array ranges over modular types. Also, use the
biggest bitvector type possible for computation. As before, if the domain
is EW_Prog and the computation may overflow (on arrays indexed by 64 bits
bitvectors) the compuation is done on mathematical integers.

* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): Only use bitvectors for the computation if the array
 ranges over modular types.

* tests/P921-026__length_bv/test.out
* tests/R305-028__counterexample/test.out
Minor update of output.

* tests/SA16-067__length
Enable test.

Change-Id: I321d32c73a8abb01897341c4f537128f4f2284d9