SA30-017 do not ignore previous writes in initialization checks
In checks for accesses to uninitialized data we attempted to distinguish
between "is uninitialized" and "might be uninitialized" by detecting cycles
in CFG, but this didn't quite work, especially for arrays and when inlining
for proof. Also, the decision was depending on the order of iteration over
hashed sets, so the test results were spuriously affected, e.g. by adding
dummy vertices to CFGs.

* gnat2why/flow/flow-analysis.adb
(Might_Be_Initialized): do not detect cycles; instead, conservatively decide
 in favour of "might be initialized".

* testsuite/gnatprove/tests/K607-022__unconstrained_array/test.out
Update; the "might" message that we got now is actually an improvement.

* testsuite/gnatprove/tests/M514-009__flow_array_init_in_loop/test.out
* testsuite/gnatprove/tests/M611-052__discr/test.out
* testsuite/gnatprove/tests/M906-011__flow_detect_conditional_...
* testsuite/gnatprove/tests/N424-011__flow_error_placement/test.out
* testsuite/gnatprove/tests/O428-055__formal_vector/test.out
* testsuite/gnatprove/tests/O512-022__demo_not_fixed_yet/test.out
* testsuite/gnatprove/tests/O512-022__infoflow/test.out
* testsuite/gnatprove/tests/O512-022__stacks/test.out
* testsuite/gnatprove/tests/OA05-076__dynamic_inv_assert_and_cut/test.out
* testsuite/gnatprove/tests/Q424-003__if_expression/test.out
Update; these seem to be regressions, but completely acceptable.

* testsuite/gnatprove/tests/SA30-017__flow_uninitialized/test.out
Update; now we get consistent "might" messages, for both CASE and IF
versions and with inlining both enabled and disabled.

Change-Id: I27ded20975bbc556cafc6129a5437087325d8599