SA17-026 trim Depends before looking for a dependency path
Once we determine a missing "LHS => RHS" dependency clause and we look for a
path in the code that is responsible for this clause, we must trim it based
on the outputs/inputs that actually has been introduced to the flow graph.

* gnat2why/flow/flow-analysis.adb
(Prepare_Trimming): reuse code for trimming in Get_Depends to get the
 inputs/outputs that actually appear on the graph

(Strip_Proof_Ins): get globals in the enclosing subprogram, so that they
 can be reused in Prepare_Trimming.

(Check_Depends_Contract): trim down-projected Missing_In and Missing_Out
 before looking for a dependency path.

* testsuite/gnatprove/tests/SA17-026__flow_depends_path/
Enable; add output, including the dependency path.

Change-Id: I1bb3df793996ef51a2c5cf0c500639d033a37055