R803-027 support partially visible predicates
In GNATprove, predicates are handled using the predicate functions
generated by the frontend. As these functions stick together all
predicates applying to the type (from the partial/private view and
inherited), GNATprove was rejecting predicates on types whose
full view is not in SPARK.
After this commit, predicate functions are only used to get the
analyzed and expanded expression for the predicates applying to the
type itself (not those inherited, which are added manually by climbing
the derivation tree). In addition, using the Aspect_On_Partial_View flag,
we can decide the visibility associated to the predicates of each type
in the derivation. As a result, we can now handle partially visible
type predicates, except in the corner case where two predicates are
supplied on the partial and full view of a unique type and the full view
is not visible in SPARK.

* gnat2why/spark/spark_definition.adb
(Mark_Call): Do not mark calls to predicate functions, they will be
 ignored during translation to Why3.
(Mark_Type_Entity): Types which have two predicates with different
 SPARK_Mode are not supported it in SPARK.
(SPARK_Pragma_Of_Entity): Factorize treatment of declarations and reuse
 the code for predicate pragmas. The applicable pragma of predicate
 functions now depends on the position of the associated aspect or pragma.

* gnat2why/spark/spark_util-types.adb
(Find_Predicate_Item): Go over a list of representation items to search
 for a predicate pragma or aspect applying to a given type.

* gnat2why/why/gnat2why-expr.adb
(Compute_Dynamic_Invariant): As calls to inherited predicates are ingored,
 we need to only ignore predicates specified on the type itself, and not
 those inherited, when Top_Predicate is False.
(Variables_In_Dynamic_Predicate)
(Compute_Dynamic_Predicate): Climb the derivation tree of the type to
 collect inherited predicates (visibly) applying to the type.
(Transform_Expr): Calls to predicate functions are ignored.

* tests/P112-038__predicate_private_type_not_in_spark/test.opt
Disable test. It contains invalid Ada code.

* tests/QC05-058__predicate/test.out
Spurious check no longer emitted (we no longer explicitly convert to the
ancestor type to check inherited predicates).

* tests/QC10-002__predicate/test.out
Code is accepted now.

* tests/R803-027__bad_predicate
New test with duplicate predicates with different visibilities.

* tests/R803-027__private_predicate
New test with partially visile predicates.

Change-Id: Ia04a531e31e71e1e8d51a88e9f20697cb0882135