SB05-027 assume pledge of borrowing traversal functions in post
Since traversal functions cannot modify their parameter, the pledge
of a borrowing traversal function holds after the call. Assume it in the
postcondition to avoid having to repeat the pledge as an immediate post.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post)
(Generate_Subprogram_Program_Fun): Add the peldge to the post of
 borrowing traversal functions.

* tests/SA10-028__old_in_pledge/test.out
Remove redundant part of the post, regain proofs lost when switching to
new version of Z3.

Change-Id: I73829ec16d3ae1614ca2386886844409073ab69d