SA16-042 Fix handling of object alignment
GNATprove was incorrectly relying on the value of alignment for the type
of objects, including in some cases when the alignment was specifically
specified on the object. Now fixed by only supporting Object'Alignment
when it is specified by the user on Object (through an aspect or a
representation clause) as otherwise its value cannot be known.

* gnat2why/spark/spark_definition.adb
(Mark_Attribute_Reference): Reject unsupported cases.

* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): Fix case of object for which alignment is specified.

* gnat2why/why/spark_atree-entities.adb
* gnat2why/why/spark_atree-entities.ads
(Alignment, Known_Alignment): Functions now apply to type or object.

* testsuite/gnatprove/tests/SA16-042__alignment/test.opt
* testsuite/gnatprove/tests/SA16-042__alignment/test.out
* testsuite/gnatprove/tests/SA16-042__alignment/test.py
Test passes now.

* testsuite/gnatprove/tests/SA16-042__bad_alignment/bad_align.adb
* testsuite/gnatprove/tests/SA16-042__bad_alignment/test.out
* testsuite/gnatprove/tests/SA16-042__bad_alignment/test.py
New test for rejected cases.

* testsuite/gnatprove/tests/MA01-040__flow_attributes_pedantic_warning/
* testsuite/gnatprove/tests/P511-009__test_attributes/
Tests modified to account for newly rejected Var'Alignment.

Change-Id: I0b131cd1b98e8f9887ca30328d5c4fb778c9ae2b