SB07-016 record state refinements for packages with no SPARK_Mode
When flow expands abstract states into constituents for proof queries, it
relies on the mapping recorded in the ALI files. This mapping was only
recorded for packages whose body has SPARK_Mode => On; now it is also
recorded for packages with no SPARK_Mode on their bodies.

This is needed because when we analyse code with SPARK_Mode => Auto, e.g.
while checking Global-vs-Pre/Post contracts, we might seen both abstract
states (in global) and their constituents (in Pre/Post), so we need to know
that they are related.

* gnat2why/flow/flow_generated_globals-partial.adb
(Preanalyze_Body): call GG_Register_State_Refinement when the SPARK_Mode
 on the body (if present) is either On or None.

(Write_Contracts_To_ALI): remove outdated and out-of-place comment.

* testsuite/gnatprove/tests/SB07-016__flow_generic_family/test.out
Primary test for this TN; spurious output is gone.

* testsuite/gnatprove/tests/P820-002__protected_out/test.out
Improvement; spurious error about a missing global is gone.

Change-Id: Id9834953aa57283555b8e7da941011bea967ec03