SB07-028 reject type aspects on partial view of derived private types
It can happen that a (private) type and its Retysp do not represent the
same entity. This happens when the type is a private type deriving from
a private type as in this case the full view of the type is private and
the underlying type of this full view is another entity. In this case,
information about DIC, predicates and invariants can differ between the
type and its full view.
As we do not currently support that in proof, we reject these cases in
marking.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Reject private types with their own predicate/DIC/
 invariants if the Retysp and the private view do not share the same
 unique entity.

* tests/N711-045__flow_default_init_cond/test.out
* tests/PA12-001__DIC/test.out
Tests are rejected in marking now.

* tests/SB07-028__ignored_DIC
Enable test.

Change-Id: I4e9367178e543b45e8a5ad599944848d3ad4a8b5