SA25-031 hierarchical folding for GG in phase 2
In phase 1 we resolve generated globals in hierarchical manner to make sure
that globals coming from calls within the same package do not introduce its
abstract states but only their constituents. This was fine.

In phase 2 we were resolving generated globals in non-hierarchical manner,
because we assumed that globals coming from calls between different
compilation units will only introduce objects that are visible from
everywhere. This was clearly wrong for calls between a parent and its
private child units. There objects (and states) declared in the private
child are visible in the parent body, but not everywhere.

This patch essentially copies the hierarchical resolution machinery from
phase 1 to phase 2, which turns out to just work. The code copied to phase 2
is as well (un)documented as the code in phase 1. This machinery relies on a
program unit hierarchy, which in phase 1 is built by traversing the AST; to
have a similar hierarchy in phase 2 we now store extra into in the ALI
files.

* gnat2why/flow/flow_generated_globals.ads
(Patrial_Contracts): add Parents component that will be communicated via the
 ALI file.

* gnat2why/flow/flow_generated_globals-traversal.ads
* gnat2why/flow/flow_generated_globals-traversal.adb
(Traversal_Parents): new routine for getting the hierarchy info that will be
 written in the ALI file in phase 1.

* gnat2why/flow/flow_generated_globals-phase_1.adb
(GG_Register_Global_Info): write the hierarchy info in phase 1.

* gnat2why/flow/flow_generated_globals-phase_2.ads
(Standard_Standard): lifted from the child Visibility package, because now
 it is used both there and in the child Traversal package.

* gnat2why/flow/flow_generated_globals-phase_2-read.ads
* gnat2why/flow/flow_generated_globals-phase_2-read.adb
(Serialize): new routine for reading a list of Entity_Names.

* gnat2why/flow/flow_generated_globals-phase_2-traversal.ads
* gnat2why/flow/flow_generated_globals-phase_2-traversal.adb
(Traversal): new package that mirrors a similar one, just that for
 Entity_Names.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(GG_Read): previously we were analysing the current unit in a hierarchical
 style and then analysing entities to translate from other units; now we
 redo the analysis of everything in a hierarchical style (just like it is
 done in phase 1 for the current compilation unit).

* testsuite/gnatprove/tests/SA25-031__flow_private_child/test.out
Update; spurious check is gone.

Change-Id: Iaf377eaba35429aa7e7b5885470736cd6135812c