R803-027 use declaration to get ancestor type of subtypes
To construct the predicate applying to a type, we climb up the derivation
tree using Etype. If we are going through subtypes, we may miss an
ancestor as the Etype of a subtype is its base type. Instead, go to
the subtype declaration to find the correct ancestor.

* gnat2why/why/gnat2why-expr.adb
(Compute_Dynamic_Predicate): On subtype use the subtype declaration to
 find the correct ancestor.

* testsuite/gnatprove/tests/tokeneer
Update sessions.

Change-Id: I3129edf5f03b7a15c271329bbaa603d41599ed32