SB07-020 add an assertion that would crash earlier on this TN
The crash observed by the customer happens when adding vertices to the CFG.
We could crash much earlier (and closer to the actual problem) if we had a
predicate for type Global_Flow_Ids that represents Global contract with
Entity_Name. We didn't have this predicate, because of a GNAT bug, though we
already have similar predicates for types Global_Nodes and Global_Names,
which represent Global contracts with Entity_Ids and Entity_Names,
respectively.

The solution is to add predicate as a pragma and not an aspect. They are
semantically equivalent, but only pragma works for GNAT.

* gnat2why/flow/flow_types.ads
(Global_Flow_Ids): add a predicate just like the one we had for Global_Set.

Change-Id: Ia02ddabe4502e359194580e473b828871e02eef4