remove useless call to Down_Project when checking Initializes
When checking the Initializes contract (which acts as a postcondition for a
package), in FA.Visible_Vars we have variables from the LHS of the
Initializes contract (added at the beginning of Control_Flow_Graph.Create)
and variables introduced to the CFG up to the private part of the package
spec. All those objects are already as seen at scope of the package spec,
so there is no need to down-project them.

(no-tn-check)

* gnat2why/flow/flow-analysis.adb
(Sanity_Check_Postcondition): remove useless call to Down_Project.

Change-Id: Iae9d6791ace04e07ac909852bb2ea6419554344b