SB07-020 fix overlapping in down-projected globals
When down-projecting a partially visible proof_in abstract state we crashed
because of an overlapping with its input/output constituents. Now this is
fixed in GG_Get_Globals. A similar problem still occurs in a Down_Project
that is used by Callee_Globals, but this should be fixed on a dedicated TN
with an appropriate regression test.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(GG_Get_Globals): avoid overlapping between proof_ins and inputs/outputs.
(Callee_Globals): add a ??? comment about a similar problem.

* testsuite/gnatprove/tests/SB07-020__flow_partial/
Enable; output is as expected.

Change-Id: I3cd18cd2adc54cc390c974277bd56444c32b6994