avoid redundant condition when dispatching on No_Return calls in Why
In Why generation the "No_Return" selector is only generated for error
signalling procedures (in Generate_Subprogram_Program_Fun). When we choose
this selector for a call to Why_Subp_Has_Precondition, we already checked
that we are processing an error signalling procedure (in
Transform_Statement_Or_Declaration).

(no-tn-check)

* gnat2why/why/gnat2why-expr.adb
(Why_Subp_Has_Precondition): move redundant condition to Pre contract.

Change-Id: Ib271a65df016050db34c0e6eddff9c877e83a0a0