QB07-016 ignore attributes 'Version and 'Body_Version in flow
Test for this TN was submitted as disabled, it was crashing in proof. In the
meantime it started crashing in flow, because we added an assertion to
sanity-check that we never see references to procedures in expressions,
which is clearly not the case for something like "My_Proc'Version".

With this patch flow analysis will ignore 'Version and 'Body_Version
attributes, so the code can at least be analysed with --mode=flow and will
crash in proof just like before. (This problem was noticed while casually
looking the continuous builder report.)

* gnat2why/flow/flow_utility.adb
(Do_Attribute_Reference): ignore attributes 'Version and 'Body_Version.

Change-Id: Ifa233a8b164934772ce6f45f0aea1cde9c8a42fc