remove redundant call to Warn_On_Subprogram_With_No_Effect
Procedure Warn_On_Subprogram_With_No_Effect was called both from
Find_Ineffective_Imports_And_Unused_Objects and from
Find_Ineffective_Statements, sometimes twice for a single analysed
subprogram, which was clearly inefficient and unnecessary.

Both calls were guarded by a complex conditions, but it turns out that the
condition for one of the call was a "subset" of the condition for the other,
so it could be safely removed.

(no-tn-check)

* gnat2why/flow/flow-analysis.adb
(Find_Ineffective_Imports_And_Unused_Objects): remove a redundant call to
 Warn_On_Subprogram_With_No_Effect.

Change-Id: I8ac7ad129081714773095738b231448c42729f29