SB29-042 add test with Part_Of abstract state and generic
This test used to crash, but this has been fixed by SB07-016. On that TN we
only witnessed a spurious error and here a crash, so it seems worth to have
this test in our regression testsuite as well.

* testsuite/gnatprove/tests/SB29-042__flow_part/
Add; it used to crash, but now is silently accepted, as it should.

Change-Id: I40d39d0d994e291433abb0b70c8ac343a68386f7