SC04-005 Check evaluation of the prefix of 'Old in classwide post
As classwide postconditions can be inherited in an overriding which
would weaken the classwide precondition, references to the 'Old
attribute in classwide postconditions should be checked to be free
of RTE without assuming the classwide precondition.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_LSP): Check that the evaluation of 'Old variables from
 the classwide post is free of RTE in the empty context.

* tests/SC04-005__LSP
Enable test.

* tests/tagged_stacks
Update test to avoid precondition failure in Post'Class.

Change-Id: I41120f0bdf1b49dcbba353a6dcb68bfaa7f336ad