SC05-027 preserve edge colours when removing null vertices
When simplifying the CFG by removing null vertices we were losing
information captured in their incoming edge colours. Luckily, it appears to
not impact the analysis, but still deserves to be fixed.

In particular, there seems to be a subtle interaction between edge colours,
Separate_Dead_Paths (which is "colour blind") and Is_Dead_End (which only
traverses "default colour" edges). Those two routines are responsible for
the flow analysis being slow on a reproducer reported on this test.

* gnat2why/flow/flow-control_flow_graph.adb
(Simplify_CFG): preserve incoming edge colours when removing null vertices.

Change-Id: I09eb21034fa69a9ba1e0fc64fe8ac25ad301d22c