SC05-027 avoid repeated CFG traversals
When detecting dead end in CFG we were repeatedly looking for a path from
the start to the analysed vertex, which was causing performance issues. Now
we discover all reachable vertices just once.

In an extreme case (with auto-generated code), it makes flow analysis 10x
faster; otherwise, behaviour is not affected.

* gnat2why/flow/flow-analysis.adb
(Find_Ineffective_Statements): precompute set of reachable nodes and use it
 in Is_Dead_End.

Change-Id: I0c94a34fcdcf1ea0b12593c8cd28be3892eb559a