simpler contract and body for Warn_On_Subprogram_With_No_Effect
The contract for Warn_On_Subprogram_With_No_Effect allowed it only to be
called on subprograms (i.e. procedures, functions and entries) with no
effects. However, function always has an effect (its result); likewise
entry (the protected unit which acts as an implicit parameter of mode
In_Out); so this routine could only be called on procedures, which is now
explicit in its contract.

With a similar reasoning, Is_Error_Signaling_Procedure (which really means
"with No_Return and no effects") in its body is equivalent to No_Return,
because the precondition ensures that there are no effects.

(no-tn-check)

* gnat2why/flow/flow-analysis.adb
(Warn_On_Subprogram_With_No_Effect): Stronger contract; simpler body.

Change-Id: Ieffe67f432d28ab89fabbb4dc903839859ea5570