S818-002 extend test with Might_Not_Return and no Global contract
In one test we have a procedure with Might_Not_Return that writes no data
(either via a global or via an In_Out/Out parameter); we don't warn about
this lack of effects, but only because there is an explicit Global => null.

Here is a test where we get this warning, though we shouldn't, because
Might_Not_Return itself announces an "effect" of possible non-termination.

* testsuite/gnatprove/tests/S818-002__might_not_return/
Extend test; warning is spurious.

Change-Id: Ib576876e0a47bdacb8dbad49e16d49f5c7a67308