SC10-033 even more relaxed contracts for Might_Not_Return
We detect No_Return/Might_Not_Return contracts for both the callee (which is
either a procedure, function or entry) and the caller (which might be also a
task or package). We didn't have tests where the caller is task or package,
but such a scenario is clear from the current usage of
Is_Possibly_Nonreturning_Procedure, e.g. in Mark_Call.

* gnat2why/spark/spark_annotate.ads
(Has_Might_Not_Return_Annotation): Relax precondition; add postcondition.

* gnat2why/spark/spark_util-subprograms.ads
(Is_Possibly_Nonreturning_Procedure): Likewise.

Change-Id: Id3721bf8e58442ee45a83371e2a36deb4dcd7b96