SC18-010 fix comparison of Flow_Ids
Comparison of Flow_Ids, which is used by Ordered_Flow_Id_Sets container, was
broken. With this change we can declare Flow_Id_Sets as an ordered set and
flow will still work.

It didn't matter much, because this container is only used for reliable
sorting of messages.

* gnat2why/flow/flow_types.adb
("<"): Use Unique_Component, very much like in "=" operator.

Change-Id: I869bed6cc2a831def8fe6c71db8fff942162ab2e