RB20-043 add test with explicit state for memory (de)allocation
Add a spark-heap.ads unit, which can be WITH-ed by the users that want to
annotate their (de)allocating subprograms with explicit Global/Depends,
which will reference the SPARK.Heap.Dynamic_Allocation abstract state.

Also, add the necessary

* include/spark-heap.ads
* include/spark-heap.adb
The minimal files needed (we need a dummy body because an abstract state
declared in spec needs a refinement).

* include/spark_heap.gpr
Supporting infrastructure for GPRbuild.

* include/setup.sh
Supporting infrastructure for GNATstudio.

* testsuite/gnatprove/tests/RB20-043__flow_heap/
Test to make sure that both the analysis and compilation works. The high
check about a missing Global is because the implicit & explicit abstract
states are not distinct; it is to demonstrate that the follow-up commit
works as expected.

Change-Id: I8d88b389c4d9e76b439ae46b061a61b2311c93a2