SC18-010 improve hashing of Flow_Ids
The Hash routine for the Flow_Id record type was only using its Variant
component when Kind was Record_Field (so when the value represented a record
field). This seems to be an oversight; now the Variant component is also
used when Kind is Direct_Mapping and Magic_String (so when the value
represents an entire object and a global from other compilation unit,
respectively).

Also, the arbitrary constants used to compute the hash are now bigger
(though still prime), because we were getting lots of collisions in the
testsuite (and now only a few), as documented in the new comment.

This change doesn't appear to affect the performance, but still feels like
the right thing to do.

* gnat2why/flow/flow_types.adb
(Hash): Use the Flow_Id.Variant component for Direct_Mapping and
 Magic_String, just like it was used for Record_Field; increase and
 document arbitrary numbers.

Change-Id: Iffc44be96f2c5035c49082a05e12cf5ab26be57f